Address PR comments - Rename view component and add previews - add visually hidden text - rename variables in test files - update radio button views - refactor string look up into translation file
