+++ app/views/support/providers/edit.html.erb
@@ -8,10 +8,11 @@
 
     <%= f.govuk_text_field :provider_name, label: { text: "Provider name", size: "m" }, width: 20 %>
 
-      <%= f.govuk_radio_buttons_fieldset :provider_type, legend: { text: "Provider type", size: "m" }, hint: { text: "Choosing SCITT or University will also mark Provider as an Accredited body" } do %>
-        <%= f.govuk_radio_button :provider_type, :scitt, label: { text: "SCITT", size: "s" }, width: 20 %>
-        <%= f.govuk_radio_button :provider_type, :lead_school, label: { text: "Lead school", size: "s" }, width: 20 %>
-        <%= f.govuk_radio_button :provider_type, :university, label: { text: "University", size: "s" }, width: 20 %>
+      <%= f.govuk_radio_buttons_fieldset :provider_type, legend: { text: "Provider type", size: "m" } do %>
+        <%= f.govuk_radio_button :provider_type, :scitt, label: { text: "SCITT", size: "s" }, hint: { text: "Marks as an accredited provider" }, width: 20 %>
+        <%= f.govuk_radio_button :provider_type, :lead_school, label: { text: "Lead school", size: "s" }, hint: { text: "Marks as not an accredited provider" }, width: 20 %>
+        <%= f.govuk_radio_divider %>
+        <%= f.govuk_radio_button :provider_type, :university, label: { text: "University", size: "s" }, hint: { text: "Marks as an accredited provider" }, width: 20 %>
       <% end %>
 
     <%= f.govuk_submit %>
