+++ app/views/support/providers/show.html.erb
@@ -1,1 +1,1 @@
-<%= render Providers::ProviderListComponent.new(provider: @provider) %>
+<%= render Providers::ProviderList::View.new(provider: @provider) %>
